Remove virtual path provider for now. Use forward slash everywhere.

+++ src/Knapsack/ModuleCache.cs
@@ -76,7 +76,10 @@ namespace Knapsack
 
         void DeleteModule(Module module)
         {
-            storage.DeleteDirectory(module.Path);
+            if (storage.DirectoryExists(module.Path))
+            {
+                storage.DeleteDirectory(module.Path);
+            }
         }
 
         void WriteManifest(ModuleContainer moduleContainer)
